Rename parse and types to xml
