+++ l5x/xml_test.go
@@ -677,13 +677,13 @@ var exampleRslogixContent = RSLogix5000Content{
 	Controller:       exampleController,
 }
 
-func TestParse(t *testing.T) {
+func TestXmlFromFile(t *testing.T) {
 	l5x, err := NewFromFile(testFilePath)
 	require.NoError(t, err)
 	require.Equal(t, &exampleRslogixContent, l5x)
 }
 
-func TestMarshall(t *testing.T) {
+func TestXmlMarshall(t *testing.T) {
 	t.Skip()
 	require.Fail(t, "Marshal tests aren't implemented yet")
 }
